requirements.txt
agency-swarm>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0
gradio>=3.50.0
slack_bolt>=1.23.0
flask>=3.1.0 
